+++ src/notification/entities/notification.entity.ts
@@ -11,6 +11,6 @@ export class Notification extends BaseModel {
   @Column({ type: 'text' })
   message: string;
 
-  @Column({ default: false })
+  @Column({ type: 'boolean', default: false })
   isRead: boolean;
 }
