+++ src/notification/notification.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Notification } from './entities/notification.entity';
-import { NotificationController } from './notification.controller';
-import { NotificationService } from './notification.service';
-import { NotificationSubscriber } from './notification.subscriber';
+import { NotificationController } from '../notification/controllers/notification.controller';
+import { NotificationService } from '../notification/services/notification.service';
+import { NotificationSubscriber } from '../notification/subscribers/notification.subscriber';
 import { AuthModule } from 'src/auth/auth.module';
 
 @Module({
