--- src/notification/notification.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { NotificationService } from './notification.service';
-
-describe('NotificationService', () => {
-  let service: NotificationService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [NotificationService],
-    }).compile();
-
-    service = module.get<NotificationService>(NotificationService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
