+++ src/notification/services/notification.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, ForbiddenException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Notification } from './entities/notification.entity';
+import { Notification } from '../entities/notification.entity';
 
 @Injectable()
 export class NotificationService {
